// src/models/entity/jvm_stack_entity.ts
/*
 * --------------------------------------------------------------------------------
 * Copyright (c) 2022-NOW(至今) 锋楪技术团队
 * Author: 锋楪技术团队 (https://www.frontleaves.com)
 *
 * 本文件包含锋楪技术团队项目的源代码，项目的所有源代码均遵循 MIT 开源许可证协议。
 * --------------------------------------------------------------------------------
 * 许可证声明：
 *
 * 版权所有 (c) 2022-2025 锋楪技术团队。保留所有权利。
 *
 * 本软件是"按原样"提供的，没有任何形式的明示或暗示的保证，包括但不限于
 * 对适销性、特定用途的适用性和非侵权性的暗示保证。在任何情况下，
 * 作者或版权持有人均不承担因软件或软件的使用或其他交易而产生的、
 * 由此引起的或以任何方式与此软件有关的任何索赔、损害或其他责任。
 *
 * 使用本软件即表示您了解此声明并同意其条款。
 *
 * 有关 MIT 许可证的更多信息，请查看项目根目录下的 LICENSE 文件或访问：
 * https://opensource.org/licenses/MIT
 * --------------------------------------------------------------------------------
 * 免责声明：
 *
 * 使用本软件的风险由用户自担。作者或版权持有人在法律允许的最大范围内，
 * 对因使用本软件内容而导致的任何直接或间接的损失不承担任何责任。
 * --------------------------------------------------------------------------------
 */

/**
 * 线程信息内部类
 *
 * ThreadInfo
 */
export type ThreadInfoEntity = {
    /**
     * 是否为守护线程
     */
    is_daemon?: boolean;
    /**
     * 线程优先级
     */
    priority?: number;
    /**
     * 线程堆栈跟踪
     */
    stack_trace?: string[];
    /**
     * 线程名称
     */
    thread_name?: string;
    /**
     * 线程状态
     */
    thread_state?: string;
}

/**
 * JvmStackDTO
 */
export type JvmStackEntity = {
    /**
     * 活动线程数
     */
    active_thread_count?: number;
    /**
     * JVM 空闲内存（字节）
     */
    free_memory?: number;
    /**
     * JVM 最大内存（字节）
     */
    max_memory?: number;
    /**
     * 系统属性
     */
    system_properties?: Map<string, string>;
    /**
     * 线程堆栈信息
     */
    thread_infos?: ThreadInfoEntity[];
    /**
     * JVM 总内存（字节）
     */
    total_memory?: number;
    /**
     * JVM 已用内存（字节）
     */
    used_memory?: number;
}